Added mob touch damage, reactor spawns, more reorganiztation
- mob touch damage
- alert stance after combat
- effects for iron body, invincibility
- reactor spawns
- extracted combat related things from stage to combat class
- rewrote data module
- moved character look related data classes into character/look
- separated equip data from clothing (sprites)
- moved spawn code into individual mapobject classes
- added modifications to color to drawargument
- moved login information from session into the login uis
- added party operation packets but they are not used yet
- moved some classes from util folder to new template folder

+++ Gameplay/Camera.cpp
@@ -17,7 +17,7 @@
 //////////////////////////////////////////////////////////////////////////////
 #include "Camera.h"
 
-#include "..\Constants.h"
+#include "../Constants.h"
 
 namespace jrc
 {
@@ -84,14 +84,14 @@ namespace jrc
 		return{ shortx, shorty };
 	}
 
-	Point<int16_t> Camera::position(double alpha) const
+	Point<int16_t> Camera::position(float alpha) const
 	{
 		auto interx = static_cast<int16_t>(std::round(x.get(alpha)));
 		auto intery = static_cast<int16_t>(std::round(y.get(alpha)));
 		return{ interx, intery };
 	}
 
-	Point<double> Camera::realposition(double alpha) const
+	Point<double> Camera::realposition(float alpha) const
 	{
 		return{ x.get(alpha), y.get(alpha) };
 	}
